+++ apps/vue-app/.storybook/main.js
@@ -1,10 +1,3 @@
-
-const coverageConfig = {
-  istanbul: {
-    include: ['**/*.stories.js'],
-  },
-};
-
 /** @type { import('@storybook/vue3-vite').StorybookConfig } */
 const config = {
   stories: ['../src/**/*.mdx', '../src/**/*.stories.js'],
@@ -12,10 +5,6 @@ const config = {
     '@storybook/addon-links',
     '@storybook/addon-docs',
     '@storybook/addon-a11y',
-    {
-      name: '@storybook/addon-coverage',
-      options: coverageConfig,
-    },
   ],
   core: {
     builder: '@storybook/builder-vite',