+++ shared/sonar-services/src/api/sonar-request.js
@@ -1,4 +1,4 @@
-import SonarRequestPolyfill from '../polyfills/sonar-request'
+import SonarRequestPolyfill from '../polyfills/sonar-request.js'
 
 const SonarRequest = globalThis.SonarRequest ?? SonarRequestPolyfill