+++ shared/sonar-services/src/polyfills/sonar-i18n.js
@@ -9,8 +9,7 @@
 const IS_DEV = globalThis?.process.env.NODE_ENV === 'development'
 
 const DEFAULT_MESSAGES = {
-   
-  default_error_message: 'The request cannot be processed. Try again later.'
+  'default_error_message': 'The request cannot be processed. Try again later.'
 }
 
 const sonarQubeL10nBundle = { messages: {} }